update/ split MemberCard & adjust defaultImages
