+++ constants/defaultImages.tsx
@@ -1,0 +1,6 @@
+import { DefaultImagesType } from "./globalTypes";
+
+// 團員的朋友預設頭像
+export const defaultImagePackage: DefaultImagesType = {
+  subMemberPhoto: "/images/logo.jpg",
+};
