+++ constants/globalTypes.tsx
@@ -5,6 +5,25 @@ export type ImgType = {
   alt: string;
 };
 
+// iconsPackage
+type IconIdType =
+  | "my-following"
+  | "my-following-active"
+  | "my-groups"
+  | "my-groups-active"
+  | "profile-setting"
+  | "profile-setting-active"
+  | "my-notification"
+  | "my-notification-active"
+  | "logout"
+  | "sub-closing"
+  | "sub-opening";
+
+export type UserNavIconsType = Record<IconIdType, ImgType>;
+
+// defaultImages
+export type DefaultImagesType = Record<string, ImgType>;
+
 // AuthProvider
 export type AuthProviderProps = {
   children: ReactNode;
