+++ modules/GroupProfile/MembersList/MemberCard.tsx
@@ -1,12 +1,10 @@
-import React, { useContext } from "react";
-import Image from "next/image";
+import React from "react";
 import Link from "@/common/components/GeneralLink";
-import { memberStatusIndex } from "@/constants/wordIndexes";
-import { GroupDataContext } from "./index";
-import { MemberCardProps } from "./data";
 import ProfileImg from "@/common/components/ProfileImg";
+import { memberStatusIndex } from "@/constants/wordIndexes";
+import { MemberCardProps } from "../data";
 
-function MemberCard({ member, subNum }: MemberCardProps) {
+export default function MemberCard({ member, subNum }: MemberCardProps) {
   const { userId, userName, status, initNum } = member;
   const isSubmember = subNum > 0;
   const isSingle = initNum === 1;
@@ -47,25 +45,5 @@ function MemberCard({ member, subNum }: MemberCardProps) {
     <Link href={`/user/${userId}`} target="_blank" className="no-underline">
       <CardContent />
     </Link>
-  );
-}
-
-export default function MembersList() {
-  const { groupData, currentMemberNum } = useContext(GroupDataContext);
-  const { totalMemberNum, members } = groupData;
-
-  return (
-    <section className="p-4 md:px-3 md:py-4">
-      <h3 className="text-md leading-6 text-center">
-        目前人數：{currentMemberNum} / {totalMemberNum}
-      </h3>
-      <div className="mt-4 flex flex-col gap-4 md:gap-3">
-        {members.map((member) => {
-          return [...Array(member.initNum)].map((_, index) => (
-            <MemberCard key={index} member={member} subNum={index} />
-          ));
-        })}
-      </div>
-    </section>
   );
 }
